[Admin] User level/roles updates and adjusts

+++ routes/api.php
@@ -102,10 +102,9 @@ Route::group([
              */
             Route::apiResource("users", AdminUserController::class);
             Route::delete("/users/{user}/photo-delete", [AdminUserController::class, "photoDelete"]);
-            Route::put("/users/{user}/promote", [AdminUserController::class, "promote"]);
-            Route::put("/users/{user}/demote", [AdminUserController::class, "demote"]);
-            Route::put("/users/{user}/{role}", [AdminUserController::class, "roleUpdate"]);
-            Route::delete("/users/{user}/{role}", [AdminUserController::class, "roleDelete"]);
+            Route::put("/users/{user}/update-level",[AdminUserController::class, "updateLevel"]);
+            Route::put("/users/roles/{user}/{role}", [AdminUserController::class, "roleUpdate"]);
+            Route::delete("/users/roles/{user}/{role}", [AdminUserController::class, "roleDelete"]);
 
             /**
              * PACKAGE CONTROLLER
